chore(workstation): Consolidate workstation packages in to subfolders
All workstation related functionality is now contained under `pkg/workstation`.
