+++ pkg/pipelines/pipeline_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/windsorcli/cli/pkg/stack"
 	"github.com/windsorcli/cli/pkg/template"
 	"github.com/windsorcli/cli/pkg/tools"
-	"github.com/windsorcli/cli/pkg/virt"
+	"github.com/windsorcli/cli/pkg/workstation/virt"
 )
 
 // =============================================================================
