+++ pkg/workstation/network/darwin_network_test.go
