+++ pkg/workstation/network/mock_network.go
