+++ pkg/workstation/network/mock_network_test.go
