+++ pkg/workstation/network/network_test.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/windsorcli/cli/pkg/config"
 	"github.com/windsorcli/cli/pkg/di"
-	"github.com/windsorcli/cli/pkg/services"
 	"github.com/windsorcli/cli/pkg/shell"
-	"github.com/windsorcli/cli/pkg/ssh"
+	"github.com/windsorcli/cli/pkg/workstation/services"
+	"github.com/windsorcli/cli/pkg/workstation/ssh"
 )
 
 // =============================================================================
