+++ pkg/workstation/network/windows_network.go
