+++ pkg/workstation/network/windows_network_test.go
