+++ pkg/workstation/services/localstack_service_test.go
