+++ pkg/workstation/services/registry_service.go
