+++ pkg/workstation/services/registry_service_test.go
