+++ pkg/workstation/ssh/client_test.go
