+++ pkg/workstation/ssh/mock_client.go
