+++ pkg/workstation/ssh/mock_client_test.go
