+++ pkg/workstation/virt/virt_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/shirou/gopsutil/mem"
 	"github.com/windsorcli/cli/pkg/config"
 	"github.com/windsorcli/cli/pkg/di"
-	"github.com/windsorcli/cli/pkg/services"
+	"github.com/windsorcli/cli/pkg/workstation/services"
 	"github.com/windsorcli/cli/pkg/shell"
 )
